+++ infoblox/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = 'Equifax CIA'
 __email__ = 'cia@equifax.com'
-__version__ = '1.5.1'
+__version__ = '1.6.0'